Updated recall/change due date letters to have "will" instead of "may" re: accruing fines
Note: I changed the AFN language too, for now, because the AFN language in this letter is already not in line with AFN standard at this point anyway (i.e., our English AFN letter is already identical to Carleton's internal letter instead of the OCUL prescription). We will have to review our AFN letters, though, to make sure we have the thing we actually want there.